chore(playground): split to multiplatform sub-module
