+++ settings.gradle
@@ -14,4 +14,4 @@ apply from: "build-tools/dependency-management.gradle"
 rootProject.name = "aditogether"
 
 include 'detekt:rules'
-include 'playground'
+include 'playground:multiplatform'
